+++ webview-ui/src/components/settings/sections/TerminalSettingsSection.tsx
@@ -1,4 +1,3 @@
-import { StringRequest } from "@shared/proto/cline/common"
 import { UpdateTerminalConnectionTimeoutResponse } from "@shared/proto/index.cline"
 import { VSCodeCheckbox, VSCodeDropdown, VSCodeOption, VSCodeTextField } from "@vscode/webview-ui-toolkit/react"
 import React, { useState } from "react"
@@ -66,12 +65,8 @@ export const TerminalSettingsSection: React.FC<TerminalSettingsSectionProps> = (
 		const target = event.target as HTMLSelectElement
 		const profileId = target.value
 
-		// Save immediately - the backend will call postStateToWebview() to update our state
-		StateServiceClient.updateDefaultTerminalProfile({
-			value: profileId || "default",
-		} as StringRequest).catch((error) => {
-			console.error("Failed to update default terminal profile:", error)
-		})
+		// Save immediately using the consolidated updateSettings approach
+		updateSetting("defaultTerminalProfile", profileId || "default")
 	}
 
 	const profilesToShow = availableTerminalProfiles
